+++ lstchain/io/__init__.py
@@ -1,43 +1,38 @@
 from .config import get_standard_config, replace_config, read_configuration_file
-from .lstcontainers import (
-    DL1ParametersContainer,
-    DispContainer,
-)
+
 from .io import (
-    get_dataset_keys,
-    auto_merge_h5files,
-    smart_merge_h5files,
-    write_simtel_energy_histogram,
-    write_mcheader,
-    write_array_info,
-    write_dl2_dataframe,
-    global_metadata,
     add_global_metadata,
-    write_metadata,
-    write_subarray_tables,
+    auto_merge_h5files,
+    get_dataset_keys,
+    global_metadata,
     read_simu_info_merged_hdf5,
-    write_calibration_data
+    smart_merge_h5files,
+    write_array_info,
+    write_calibration_data,
+    write_dl2_dataframe,
+    write_mcheader,
+    write_metadata,
+    write_simtel_energy_histogram,
+    write_subarray_tables,
 )
 
 standard_config = get_standard_config()
 
 __all__ = [
-    'get_standard_config',
-    'replace_config',
-    'read_configuration_file',
-    'DL1ParametersContainer',
-    'DispContainer',
-    'get_dataset_keys',
-    'auto_merge_h5files',
-    'smart_merge_h5files',
-    'write_simtel_energy_histogram',
-    'write_mcheader',
-    'write_array_info',
-    'write_dl2_dataframe',
-    'global_metadata',
     'add_global_metadata',
-    'write_metadata',
-    'write_subarray_tables',
+    'auto_merge_h5files',
+    'get_dataset_keys',
+    'get_standard_config',
+    'global_metadata',
+    'read_configuration_file',
     'read_simu_info_merged_hdf5',
-    'write_calibration_data'
+    'replace_config',
+    'smart_merge_h5files',
+    'write_array_info',
+    'write_calibration_data',
+    'write_dl2_dataframe',
+    'write_mcheader',
+    'write_metadata',
+    'write_simtel_energy_histogram',
+    'write_subarray_tables',
 ]
