+++ lstchain/reco/r0_to_dl1.py
@@ -41,13 +41,10 @@ from ..datachecks.dl1_checker import check_dl1
 from ..image.muon import analyze_muon_event, tag_pix_thr
 from ..image.muon import create_muon_table, fill_muon_event
 from ..io import (
-    DL1ParametersContainer,
-    replace_config,
-    standard_config,
-)
-from ..io import (
     add_global_metadata,
     global_metadata,
+    replace_config,
+    standard_config,
     write_array_info,
     write_calibration_data,
     write_mcheader,
@@ -55,9 +52,15 @@ from ..io import (
     write_simtel_energy_histogram,
     write_subarray_tables,
 )
-from ..io.io import add_column_table
-from ..io.io import write_array_info_08
-from ..io.lstcontainers import ExtraImageInfo, DL1MonitoringEventIndexContainer
+from ..io.io import (
+    add_column_table,
+    write_array_info_08,
+)
+from ..lstcontainers import (
+    DL1MonitoringEventIndexContainer,
+    DL1ParametersContainer,
+    ExtraImageInfo,
+)
 from ..paths import parse_r0_filename, run_to_dl1_filename, r0_to_dl1_filename
 from ..pointing import PointingPosition
